+++ src/CControl/Sources/Miscellaneous/cut.c
@@ -16,15 +16,15 @@
  * If you have a matrix A 5 x 6 and you want to cut the values from A to matrix B with size 3 x 3 and you want to start at 0,0 and end at 3,3
  * Code: cut(A, 5, 6, B, 0, 2, 0, 2); // Because indexing from zero
  */
-void cut(float A[], uint16_t column, float B[], uint16_t start_row, uint16_t stop_row, uint16_t start_column, uint16_t stop_column) {
+void cut(float A[], size_t column, float B[], size_t start_row, size_t stop_row, size_t start_column, size_t stop_column) {
 
 	float *data = A + start_row * column + start_column;
 
 	/* Create the output */
-	uint16_t out_columns = stop_column - start_column + 1;
+	size_t out_columns = stop_column - start_column + 1;
 
 	/* Instead of having two for loops, we just copy the whole row at once. */
-	uint16_t i;
+	size_t i;
 	for (i = start_row; i < stop_row + 1; i++) {
 		memcpy(B, data, sizeof(float) * out_columns);
 		B += out_columns;
