+++ src/CControl/Sources/Miscellaneous/insert.c
@@ -12,12 +12,12 @@
  * The variables startRow_b and startColumn_b describe which row and column we should to insert first element of A into B
  * Notice that start positions are indexed from zero
  */
-void insert(float A[], float B[], uint16_t row_a, uint16_t column_a, uint16_t column_b, uint16_t startRow_b, uint16_t startColumn_b){
+void insert(float A[], float B[], size_t row_a, size_t column_a, size_t column_b, size_t startRow_b, size_t startColumn_b){
 	/* Initial */
 	B += startRow_b*column_b + startColumn_b; /* Now we are standing on position startRow_b x startColumn_b of row_b x column_b */
 
 	/* We start at startRow_b and end at row_a, then we have inserted all rows from A into B */
-	uint16_t i;
+	size_t i;
 	for(i = 0; i < row_a; i++){
 		memcpy(B, A, column_a*sizeof(float)); /* Insert the complete row from A into B */
 		A += column_a;
