+++ src/CControl/Sources/Statistics/cov.c
@@ -13,11 +13,11 @@
    * y[L] Vector with values
    * L = Length of vector x and y
    */
-float cov(float x[], float y[], uint16_t length) {
+float cov(float x[], float y[], size_t length) {
 	float x_mu = mean(x, length);
 	float y_mu = mean(y, length);
 	float sigma = 0.0f;
-	uint16_t i;
+	size_t i;
 	for (i = 0; i < length; i++) {
 		sigma += (x[i] - x_mu)*(y[i] - y_mu);
 	}
